fixed pagination isue

+++ main.php
@@ -64,18 +64,19 @@ if (!class_exists('KHMYCLASS')) {
             if ($this->isdataenabled === '1') {
                 add_action('wpforms_process_entry_save', array($this, 'process_entry'), 10, 4);
             }
-            add_action('wp_ajax_update_form_values', array($this, 'update_form_values'));
-            add_action('wp_ajax_nopriv_update_form_values', array($this, 'update_form_values'));
-            add_action('wp_ajax_get_form_values', array($this, 'get_form_values'));
-            add_action('wp_ajax_nopriv_get_form_values', array($this, 'get_form_values'));
-            add_action('wp_ajax_delete_form_row', array($this, 'delete_form_row'));
-            add_action('wp_ajax_nopriv_delete_form_row', array($this, 'delete_form_row'));
+
 
             // Other actions
             // add_filter('login_redirect', array($this, 'custom_login_redirect'), 10, 3);
             add_action('wp_login', array($this, 'redirect_to_saved_url'));
 
+
+
+
         }
+
+
+
 
 
 
@@ -84,9 +85,9 @@ if (!class_exists('KHMYCLASS')) {
         {
             $saved_url = get_option('saved_url');
             if (!empty($saved_url)) {
-                //$login_url = wp_login_url(add_query_arg('redirect', 'specific_value'));
                 $redirect_param = isset($_GET['redirect']) ? $_GET['redirect'] : '';
-                if ($redirect_param === 'specific_value') {
+                $paged_param = isset($_GET['paged']) ? $_GET['paged'] : ''; // Check for 'paged' parameter
+                if ($redirect_param === 'specific_value' && empty($paged_param)) {
                     delete_option('saved_url');
                     wp_redirect($saved_url);
                     exit;
@@ -179,6 +180,9 @@ if (!class_exists('KHMYCLASS')) {
             include_once KHFORM_PATH . 'Inc/KHSettings.php';
             include_once KHFORM_PATH . 'Inc/KHPDF.php';
             include_once KHFORM_PATH . 'Inc/KHdb.php';
+            include_once KHFORM_PATH . 'Inc/AjaxClass.php';
+            include_once KHFORM_PATH . 'Inc/KHwidget.php';
+
         }
 
         /**
@@ -396,13 +400,7 @@ if (!class_exists('KHMYCLASS')) {
         }
 
 
-        /**
-         * display form values shortcode
-         *
-         * @since 1.0.0
-         */
-        //
-        function display_form_values_shortcode($atts)
+        function display_form_values_shortcode2($atts)
         {
             global $wpdb;
             $atts = shortcode_atts(
@@ -412,7 +410,14 @@ if (!class_exists('KHMYCLASS')) {
                 $atts
             );
 
+            // Pagination logic
+            $current_page = max(1, get_query_var('paged'));
 
+            $items_per_page = get_option('number_id_setting');
+            if (empty($items_per_page)) {
+                $items_per_page = 10;
+            }
+            $offset = ($current_page - 1) * $items_per_page;
 
             // see if user do not have authorization 
             if (!current_user_can('manage_options')) {
@@ -430,16 +435,152 @@ if (!class_exists('KHMYCLASS')) {
                 if (!empty($atts['id'])) {
                     $formbyid = $atts['id'];
                 } else {
-                    $formbyid = $this->mydb->retrieve_form_id();
+                    $formbyid = KHdb::getInstance()->retrieve_form_id();
 
                 }
 
                 error_log('display the changed form id' . $formbyid);
                 // retrieve form values
-                $form_values = $this->mydb->retrieve_form_values($formbyid);
+                $form_values = KHdb::getInstance()->retrieve_form_values($formbyid, $offset, $items_per_page);
 
                 //Check if there is at least one entry
-                if ($this->mydb->is_table_empty() === true) {
+                if (KHdb::getInstance()->is_table_empty() === true) {
+                    ob_start();
+
+                    echo '<div style="text-align: center; color: red;">No data available! Please add entries to your form and try again.';
+                    echo ' <a style="text-align: center; color: black;" href="' . admin_url('admin.php?page=khwplist.php') . '">Settings
+                    DB</a></div>';
+
+                    return ob_get_clean();
+
+                } else {
+                    ob_start();
+
+                    //include edit-form file
+                    include_once KHFORM_PATH . 'Inc/html/edit_popup.php';
+                    // Start table
+                    echo '<table style="border: 1px solid black;">';
+
+                    // Table header
+                    echo '<tr>';
+                    echo '<th>ID</th>';
+                    echo '<th>Form ID</th>';
+                    echo '<th>Data</th>';
+                    echo '</tr>';
+
+                    foreach ($form_values as $form_value) {
+                        $form_id = intval($form_value['form_id']);
+                        $id = intval($form_value['id']);
+
+                        // Table row
+                        echo '<tr style="border: 1px solid black;" >';
+                        echo '<td style="border: 1px solid black;  padding: 10px; text-align: center;">' . $id . '</td>';
+                        echo '<td style="border: 1px solid black;  padding: 10px; text-align: center;">' . $form_id . '</td>';
+                        echo '<td style="border: 1px solid black;">';
+
+                        // Table data
+                        foreach ($form_value['data'] as $key => $value) {
+                            if (empty($value)) {
+                                continue;
+                            }
+
+                            echo '<div>';
+                            echo '<span>' . $key . ': </span>';
+                            echo '<span>' . $value . ' </span>';
+
+                            echo '</div>';
+                        }
+
+                        echo '<button class="deletebtn" data-form-id="' . esc_attr($id) . '" data-nonce="' . wp_create_nonce('ajax-nonce') . '">
+                        <i class="fas fa-trash"></i></button>';
+                        //<button class="delete-btn" data-form-id="' . esc_attr($id) . '"
+                        //data-nonce="' . wp_create_nonce('ajax-nonce') . '">
+                        //<i class="fas fa-trash"></i></button>
+                        echo '<button class="editbtn" 
+                        data-form-id="' . esc_attr($form_id) . '" data-id="' . esc_attr($id) . '"><i
+                        class="fas fa-edit"></i></button>';
+
+                        echo '</td>';
+                        echo '</tr>';
+                    }
+
+                    // End table
+                    echo '</table>';
+
+                    // Pagination link
+                    echo paginate_links(
+                        array(
+                            'base' => esc_url(add_query_arg('paged', '%#%')),
+                            'format' => '',
+                            'prev_text' => __('&laquo; Previous'),
+                            'next_text' => __('Next &raquo;'),
+                            'total' => ceil($wpdb->get_var("SELECT COUNT(id) FROM $this->table_name ") / $items_per_page),
+                            'current' => $current_page,
+                        )
+                    );
+
+                    echo '<br>';
+                    echo '<button style="background:' . $this->exportbgcolor . ';" class="export-btn"><i class="fas fa-download"></i> Export as CSV</button>';
+                    echo '<button style="background:' . $this->exportbgcolor . ';" class="export-btn-pdf"><i class="fas fa-download"></i> Export as PDF</button>';
+
+                    return ob_get_clean();
+                }
+            }
+        }
+
+
+        /**
+         * display form values shortcode
+         *
+         * @since 1.0.0
+         */
+        //
+        function display_form_values_shortcode($atts)
+        {
+            global $wpdb;
+            $atts = shortcode_atts(
+                array(
+                    'id' => '',
+                ),
+                $atts
+            );
+
+            // Pagination logic
+            $current_page = max(1, get_query_var('paged'));
+
+            $items_per_page = get_option('number_id_setting');
+            error_log('items_per_page' . $items_per_page);
+            if (empty($items_per_page)) {
+                $items_per_page = 10;
+            }
+            $offset = ($current_page - 1) * $items_per_page;
+
+            // see if user do not have authorization 
+            if (!current_user_can('manage_options')) {
+                // Assuming you have a link that takes users to the login page, you can add the referer URL as a query parameter.
+
+                ob_start();
+
+                echo '<div style="text-align: center; color: red;">You are not authorized to access this page. <a href="' . wp_login_url(add_query_arg('redirect', 'wpfurl')) . '">Login</a></div>';                //echo 'login: ' . wp_login_url();
+
+                return ob_get_clean();
+
+            } else {
+
+                //get the form id
+                // if (!empty($atts['id'])) {
+                //   $formbyid = $atts['id'];
+                //} else {
+                $formbyid = KHdb::getInstance()->retrieve_form_id();
+
+                //}
+
+                error_log('display the changed form id' . $formbyid);
+                // retrieve form values
+                $form_values = KHdb::getInstance()->retrieve_form_values($formbyid, $offset, $items_per_page);
+
+                //Check if there is at least one entry
+                if (KHdb::getInstance()->is_table_empty() === true) {
                     ob_start();
 
                     echo '<div style="text-align: center; color: red;">No data available! Please add etries to your form and try again.';
@@ -468,7 +609,7 @@ if (!class_exists('KHMYCLASS')) {
 
                     if ($form_values) {
                         echo '<div class="container">';
-                        echo 'Number of forms submitted: ' . $this->mydb->count_items($formbyid);
+                        echo 'Number of forms submitted: ' . KHdb::getInstance()->count_items($formbyid);
                         if (!empty($formbyid)) {
                             echo '<br> Default form id: ' . (($formbyid === '1') ? 'Show all forms' : $formbyid);
                         }
@@ -516,6 +657,19 @@ if (!class_exists('KHMYCLASS')) {
                             echo '</div>';
                         }
 
+                        //Pagination link
+                        echo paginate_links(
+                            array(
+                                'base' => esc_url(add_query_arg('paged', '%#%')),
+                                'format' => '',
+                                'prev_text' => __('&laquo; Previous'),
+                                'next_text' => __('Next &raquo;'),
+                                'total' => ceil($wpdb->get_var("SELECT COUNT(id) FROM $this->table_name ") / $items_per_page),
+                                'current' => $current_page,
+                            )
+                        );
+
+                        echo '<br>';
                         echo '<button style="background:' . $this->exportbgcolor . ';" class="export-btn"><i class="fas fa-download"></i> Export as CSV</button>';
                         echo '<button style="background:' . $this->exportbgcolor . ';" class="export-btn-pdf"><i class="fas fa-download"></i> Export as PDF</button>';
 
@@ -536,6 +690,7 @@ if (!class_exists('KHMYCLASS')) {
          */
         function process_entry($fields, $entry, $form_data, $entry_id)
         {
+
 
             global $wpdb;
